here we set all the auth verification, for login and CRUD

+++ app/Http/Controllers/Api/V2/PostController.php
@@ -8,7 +8,9 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use Illuminate\Container\Attributes\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
 
 class PostController extends Controller
 {
@@ -17,7 +19,9 @@ class PostController extends Controller
      */
     public function index()
     {
-         return PostResource::collection(Post::all());
+        $user = request()->user();
+        $post = $user->posts()->with('author')->paginate();
+         return PostResource::collection($post);
     }
 
     /**
@@ -26,7 +30,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
         $post = Post::create($data);
         return new PostResource($post);
     }
@@ -39,6 +43,13 @@ class PostController extends Controller
     */
     public function show(Post $post)
     {
+        // $user = request()->user();
+        // if($user->id != $post->author_id){
+        //     abort(403, 'Access forbidden');
+        // }
+        
+        //   THIS LINE MAKES THE SAME AS THE IF CLAUSULE ABOVE BUT IS A SHORT WAY
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access forbidden');
         // $post = Post::findOrFail($id);
         return response()->json( new PostResource($post));
 
@@ -49,7 +60,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $data = $request->validated([
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access forbidden');
+        $data = $request->validate([
             'title' => 'required|string|min:2',
             'body' => ['required', 'string', 'min:2']
         ]);
@@ -62,6 +74,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access forbidden');
         $post->delete();
         return response()->noContent();
     }
